FortmulaRESTServer/requirements/common.txt
Django==2.1
djangorestframework==3.8
psycopg2==2.7
django-cors-headers